move to custom view action
