+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/fragment/review/AddRestaurantReviewFragmentTest.kt
@@ -1,18 +1,13 @@
 package com.github.sdp.ratemyepfl.fragment.review
 
 import android.content.Intent
-import android.view.View
-import android.widget.RatingBar
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.closeSoftKeyboard
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.UiController
-import androidx.test.espresso.ViewAction
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.action.ViewActions.typeText
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import com.github.sdp.ratemyepfl.R
@@ -20,9 +15,9 @@ import com.github.sdp.ratemyepfl.activity.ReviewActivity
 import com.github.sdp.ratemyepfl.auth.FakeConnectedUser
 import com.github.sdp.ratemyepfl.model.review.ReviewRating
 import com.github.sdp.ratemyepfl.utils.CustomViewActions
+import com.github.sdp.ratemyepfl.utils.CustomViewActions.RatingAction.performSetRating
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
-import org.hamcrest.Matcher
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
@@ -119,26 +114,5 @@ class AddRestaurantReviewFragmentTest {
         onView(withId(R.id.doneButton)).perform(click())
         onView(withId(R.id.addReviewComment)).check(matches(withText(comment)))
         onView(withId(R.id.addReviewTitle)).check(matches(withText(title)))
-    }
-
-    companion object {
-        private fun performSetRating(value: Float) = object : ViewAction {
-            override fun getConstraints(): Matcher<View> {
-                return ViewMatchers.isAssignableFrom(RatingBar::class.java)
-            }
-
-            override fun getDescription(): String {
-                return "Custom view action to set rating."
-            }
-
-            override fun perform(uiController: UiController?, view: View?) {
-                val ratingBar = view as RatingBar
-                ratingBar.rating = value
-            }
-
-        }
-
-        fun performSetRating(rating: ReviewRating) =
-            performSetRating(rating.rating.toFloat())
     }
 }
